[bootstrap-agent] Refactor: Inner -> SprocketsServer
This is a nearly-trivial refactor: move the `Inner` raw TCP server out
of `bootstrap/server.rs` into its own module. This is building up to #3827.
